decouple compression
