+++ lib/segment/src/index/field_index/full_text_index/compressed_posting/compressed_chunks_reader.rs
@@ -1,13 +1,9 @@
 use bitpacking::BitPacker;
 use common::types::PointOffsetType;
 
-use crate::index::field_index::full_text_index::compressed_posting::compressed_common::BitPackerImpl;
-
-#[derive(Clone, Debug, Default)]
-pub struct CompressedPostingChunk {
-    pub initial: PointOffsetType,
-    pub offset: u32,
-}
+use crate::index::field_index::full_text_index::compressed_posting::compressed_common::{
+    BitPackerImpl, CompressedPostingChunk,
+};
 
 pub struct ChunkReader<'a> {
     data: &'a [u8],
